requirements.txt
streamlit==1.38.0
pandas==2.2.3
langchain-core==0.3.12
langchain-community==0.3.3
faiss-cpu==1.9.0
langchain-huggingface==0.1.0
google-generativeai==0.8.3